+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## v0.5.4
+- Add ability to query related IOCs for Threat Actors
+- Add new assessment reporting dashboard
+
 ## v0.5.3
 - Minor style changes, cleanup & bugfixes